pyproject.toml
[tool.poetry]
name = "django-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
gql = "^3.5.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
requests-toolbelt = "^1.0.0"
selenium = "^4.27.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
loguru = "^0.7.3"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
